Adds aliases for all unrestricted special pages
Made the decision to keep them all noun groups, rather than have some be imperative verbs or full sentences, since a page is a thing, not an action. The one I struggled with the most was "Diff". Went with "ante ante", since "ante" is used to mean "edit" in a number of places and I'm calling "ComparePages" "ante lipu".